packages/common/pyproject.toml
[project]
name = "snaptrash-common"
version = "0.1.0"
description = "Shared lib: Databricks client, schemas, table DDL"
requires-python = ">=3.11"
dependencies = [
    "databricks-sdk>=0.40.0",
    "pydantic>=2.7",
    "pydantic-settings>=2.4",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/snaptrash_common"]
